(Re)factor fetcher out to its own file
