+++ src/CMakeLists.txt
@@ -51,6 +51,7 @@ set(
     rdkafka_mock_handlers.c
     rdkafka_mock_cgrp.c
     rdkafka_error.c
+    rdkafka_fetcher.c
     rdlist.c
     rdlog.c
     rdmurmur2.c
