+++ src/Makefile
@@ -55,7 +55,7 @@ SRCS=		rdkafka.c rdkafka_broker.c rdkafka_msg.c rdkafka_topic.c \
 		rdkafka_txnmgr.c rdkafka_coord.c \
 		rdvarint.c rdbuf.c rdmap.c rdunittest.c \
 		rdkafka_mock.c rdkafka_mock_handlers.c rdkafka_mock_cgrp.c \
-		rdkafka_error.c \
+		rdkafka_error.c rdkafka_fetcher.c \
 		$(SRCS_y)
 
 HDRS=		rdkafka.h rdkafka_mock.h
