+++ src/rdkafka_partition.c
@@ -31,6 +31,7 @@
 #include "rdkafka_request.h"
 #include "rdkafka_offset.h"
 #include "rdkafka_partition.h"
+#include "rdkafka_fetcher.h"
 #include "rdregex.h"
 #include "rdports.h" /* rd_qsort_r() */
 
@@ -1927,171 +1928,6 @@ static void rd_kafka_toppar_pause_resume(rd_kafka_toppar_t *rktp,
         }
 }
 
-
-
-/**
- * @brief Decide whether this toppar should be on the fetch list or not.
- *
- * Also:
- *  - update toppar's op version (for broker thread's copy)
- *  - finalize statistics (move rktp_offsets to rktp_offsets_fin)
- *
- * @returns the partition's Fetch backoff timestamp, or 0 if no backoff.
- *
- * @locality broker thread
- * @locks none
- */
-rd_ts_t rd_kafka_toppar_fetch_decide(rd_kafka_toppar_t *rktp,
-                                     rd_kafka_broker_t *rkb,
-                                     int force_remove) {
-        int should_fetch   = 1;
-        const char *reason = "";
-        int32_t version;
-        rd_ts_t ts_backoff      = 0;
-        rd_bool_t lease_expired = rd_false;
-
-        rd_kafka_toppar_lock(rktp);
-
-        /* Check for preferred replica lease expiry */
-        lease_expired = rktp->rktp_leader_id != rktp->rktp_broker_id &&
-                        rd_interval(&rktp->rktp_lease_intvl,
-                                    5 * 60 * 1000 * 1000 /*5 minutes*/, 0) > 0;
-        if (lease_expired) {
-                /* delete_to_leader() requires no locks to be held */
-                rd_kafka_toppar_unlock(rktp);
-                rd_kafka_toppar_delegate_to_leader(rktp);
-                rd_kafka_toppar_lock(rktp);
-
-                reason       = "preferred replica lease expired";
-                should_fetch = 0;
-                goto done;
-        }
-
-        /* Forced removal from fetch list */
-        if (unlikely(force_remove)) {
-                reason       = "forced removal";
-                should_fetch = 0;
-                goto done;
-        }
-
-        if (unlikely((rktp->rktp_flags & RD_KAFKA_TOPPAR_F_REMOVE) != 0)) {
-                reason       = "partition removed";
-                should_fetch = 0;
-                goto done;
-        }
-
-        /* Skip toppars not in active fetch state */
-        if (rktp->rktp_fetch_state != RD_KAFKA_TOPPAR_FETCH_ACTIVE) {
-                reason       = "not in active fetch state";
-                should_fetch = 0;
-                goto done;
-        }
-
-        /* Update broker thread's fetch op version */
-        version = rktp->rktp_op_version;
-        if (version > rktp->rktp_fetch_version ||
-            rktp->rktp_next_offset != rktp->rktp_last_next_offset ||
-            rktp->rktp_offsets.fetch_offset == RD_KAFKA_OFFSET_INVALID) {
-                /* New version barrier, something was modified from the
-                 * control plane. Reset and start over.
-                 * Alternatively only the next_offset changed but not the
-                 * barrier, which is the case when automatically triggering
-                 * offset.reset (such as on PARTITION_EOF or
-                 * OFFSET_OUT_OF_RANGE). */
-
-                rd_kafka_dbg(rktp->rktp_rkt->rkt_rk, TOPIC, "FETCHDEC",
-                             "Topic %s [%" PRId32
-                             "]: fetch decide: "
-                             "updating to version %d (was %d) at "
-                             "offset %" PRId64 " (was %" PRId64 ")",
-                             rktp->rktp_rkt->rkt_topic->str,
-                             rktp->rktp_partition, version,
-                             rktp->rktp_fetch_version, rktp->rktp_next_offset,
-                             rktp->rktp_offsets.fetch_offset);
-
-                rd_kafka_offset_stats_reset(&rktp->rktp_offsets);
-
-                /* New start offset */
-                rktp->rktp_offsets.fetch_offset = rktp->rktp_next_offset;
-                rktp->rktp_last_next_offset     = rktp->rktp_next_offset;
-
-                rktp->rktp_fetch_version = version;
-
-                /* Clear last error to propagate new fetch
-                 * errors if encountered. */
-                rktp->rktp_last_error = RD_KAFKA_RESP_ERR_NO_ERROR;
-
-                rd_kafka_q_purge_toppar_version(rktp->rktp_fetchq, rktp,
-                                                version);
-        }
-
-
-        if (RD_KAFKA_TOPPAR_IS_PAUSED(rktp)) {
-                should_fetch = 0;
-                reason       = "paused";
-
-        } else if (RD_KAFKA_OFFSET_IS_LOGICAL(rktp->rktp_next_offset)) {
-                should_fetch = 0;
-                reason       = "no concrete offset";
-
-        } else if (rd_kafka_q_len(rktp->rktp_fetchq) >=
-                   rkb->rkb_rk->rk_conf.queued_min_msgs) {
-                /* Skip toppars who's local message queue is already above
-                 * the lower threshold. */
-                reason       = "queued.min.messages exceeded";
-                should_fetch = 0;
-
-        } else if ((int64_t)rd_kafka_q_size(rktp->rktp_fetchq) >=
-                   rkb->rkb_rk->rk_conf.queued_max_msg_bytes) {
-                reason       = "queued.max.messages.kbytes exceeded";
-                should_fetch = 0;
-
-        } else if (rktp->rktp_ts_fetch_backoff > rd_clock()) {
-                reason       = "fetch backed off";
-                ts_backoff   = rktp->rktp_ts_fetch_backoff;
-                should_fetch = 0;
-        }
-
-done:
-        /* Copy offset stats to finalized place holder. */
-        rktp->rktp_offsets_fin = rktp->rktp_offsets;
-
-        if (rktp->rktp_fetch != should_fetch) {
-                rd_rkb_dbg(
-                    rkb, FETCH, "FETCH",
-                    "Topic %s [%" PRId32
-                    "] in state %s at offset %s "
-                    "(%d/%d msgs, %" PRId64
-                    "/%d kb queued, "
-                    "opv %" PRId32 ") is %s%s",
-                    rktp->rktp_rkt->rkt_topic->str, rktp->rktp_partition,
-                    rd_kafka_fetch_states[rktp->rktp_fetch_state],
-                    rd_kafka_offset2str(rktp->rktp_next_offset),
-                    rd_kafka_q_len(rktp->rktp_fetchq),
-                    rkb->rkb_rk->rk_conf.queued_min_msgs,
-                    rd_kafka_q_size(rktp->rktp_fetchq) / 1024,
-                    rkb->rkb_rk->rk_conf.queued_max_msg_kbytes,
-                    rktp->rktp_fetch_version,
-                    should_fetch ? "fetchable" : "not fetchable: ", reason);
-
-                if (should_fetch) {
-                        rd_dassert(rktp->rktp_fetch_version > 0);
-                        rd_kafka_broker_active_toppar_add(
-                            rkb, rktp, *reason ? reason : "fetchable");
-                } else {
-                        rd_kafka_broker_active_toppar_del(rkb, rktp, reason);
-                }
-        }
-
-        rd_kafka_toppar_unlock(rktp);
-
-        /* Non-fetching partitions will have an
-         * indefinate backoff, unless explicitly specified. */
-        if (!should_fetch && !ts_backoff)
-                ts_backoff = RD_TS_MAX;
-
-        return ts_backoff;
-}
 
 
 /**
