+++ src/rdkafka_partition.h
@@ -524,12 +524,6 @@ void rd_kafka_toppar_fetch_stopped(rd_kafka_toppar_t *rktp,
 
 
 
-rd_ts_t rd_kafka_toppar_fetch_decide(rd_kafka_toppar_t *rktp,
-                                     rd_kafka_broker_t *rkb,
-                                     int force_remove);
-
-
-
 rd_ts_t rd_kafka_broker_consumer_toppar_serve(rd_kafka_broker_t *rkb,
                                               rd_kafka_toppar_t *rktp);
